Storage Tests and Changed tests to their specific places
[Fixed] Storage Tests
[Updated] Updated tests and separated them to their specific Contract so it's easier to track the issues, increase code readability,
