+++ Phantasma.Tests/AssemblerTests.cs
@@ -1,11 +1,11 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+
 using System;
 using System.Linq;
 using System.Text;
 using System.Numerics;
 using System.Collections.Generic;
 using System.Transactions;
-
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Phantasma.Core.Types;
 using Phantasma.Simulator;
 using Phantasma.Core.Cryptography;
@@ -25,8 +25,7 @@ namespace Phantasma.LegacyTests
     public class AssemblerTests
     {
         public const int TestGasLimit = 9999;
-
-
+        
         [TestMethod]
         public void Alias()
         {
@@ -137,426 +136,8 @@ namespace Phantasma.LegacyTests
             Assert.IsTrue(eventMessage.AsString() == message);
         }
 
-        [TestMethod]
-        public void TokenTriggers()
-        {
-            string[] scriptString;
-            //TestVM vm;
 
-
-            var owner = PhantasmaKeys.Generate();
-            var target = PhantasmaKeys.Generate();
-            var symbol = "TEST";
-            var flags = TokenFlags.Transferable | TokenFlags.Finite | TokenFlags.Fungible | TokenFlags.Divisible;
-            
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            string message = "customEvent";
-            var addressStr = Base16.Encode(owner.Address.ToByteArray());
-
-            scriptString = new string[]
-            {
-                $"alias r1, $triggerSend",
-                $"alias r2, $triggerReceive",
-                $"alias r3, $triggerBurn",
-                $"alias r4, $triggerMint",
-                $"alias r5, $currentTrigger",
-                $"alias r6, $comparisonResult",
-
-                $@"load $triggerSend, ""{TokenTrigger.OnSend}""",
-                $@"load $triggerReceive, ""{TokenTrigger.OnReceive}""",
-                $@"load $triggerBurn, ""{TokenTrigger.OnBurn}""",
-                $@"load $triggerMint, ""{TokenTrigger.OnMint}""",
-                $"pop $currentTrigger",
-
-                $"equal $triggerSend, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @sendHandler",
-
-                $"equal $triggerReceive, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @receiveHandler",
-
-                $"equal $triggerBurn, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @burnHandler",
-
-                $"equal $triggerMint, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @OnMint",
-
-                $"ret",
-
-                $"@sendHandler: ret",
-
-                $"@receiveHandler: ret",
-
-                $"@burnHandler: load r7 \"test burn handler exception\"",
-                $"throw r7",
-
-                $"@OnMint: ret",
-            };
-
-            var script = AssemblerUtils.BuildScript(scriptString, null, out var debugInfo, out var labels);
-
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, $"{symbol}Token", 1000000000, 3, flags, script, labels);
-            var tx = simulator.MintTokens(owner, owner.Address, symbol, 1000);
-            simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, symbol, 10);
-            simulator.EndBlock();
-
-            //var token = simulator.Nexus.GetTokenInfo(simulator.Nexus.RootStorage, symbol);
-            //var balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, owner.Address);
-            //Assert.IsTrue(balance == 1000);
-
-            //Assert.ThrowsException<ChainException>(() =>
-            //{
-            //    simulator.BeginBlock();
-            //    simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain, symbol, 10);
-            //    simulator.EndBlock();
-            //});
-
-            //balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, owner.Address);
-            //Assert.IsTrue(balance == 1000);
-        }
-
-        [TestMethod]
-        public void TokenTriggersEventPropagation()
-        {
-            string[] scriptString;
-            //TestVM vm;
-
-
-            var owner = PhantasmaKeys.Generate();
-            var target = PhantasmaKeys.Generate();
-            var symbol = "TEST";
-            var flags = TokenFlags.Transferable | TokenFlags.Finite | TokenFlags.Fungible | TokenFlags.Divisible;
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            string message = "customEvent";
-            var addressStr = Base16.Encode(owner.Address.ToByteArray());
-
-            scriptString = new string[]
-            {
-                $"alias r1, $triggerSend",
-                $"alias r2, $triggerReceive",
-                $"alias r3, $triggerBurn",
-                $"alias r4, $triggerMint",
-                $"alias r5, $currentTrigger",
-                $"alias r6, $comparisonResult",
-
-                $@"load $triggerSend, ""{TokenTrigger.OnSend}""",
-                $@"load $triggerReceive, ""{TokenTrigger.OnReceive}""",
-                $@"load $triggerBurn, ""{TokenTrigger.OnBurn}""",
-                $@"load $triggerMint, ""{TokenTrigger.OnMint}""",
-                $"pop $currentTrigger",
-
-                //$"equal $triggerSend, $currentTrigger, $comparisonResult",
-                //$"jmpif $comparisonResult, @sendHandler",
-
-                //$"equal $triggerReceive, $currentTrigger, $comparisonResult",
-                //$"jmpif $comparisonResult, @receiveHandler",
-
-                //$"equal $triggerBurn, $currentTrigger, $comparisonResult",
-                //$"jmpif $comparisonResult, @burnHandler",
-
-                //$"equal $triggerMint, $currentTrigger, $comparisonResult",
-                //$"jmpif $comparisonResult, @OnMint",
-
-                $"jmp @return",
-
-                $"@sendHandler: load r7 \"test send handler exception\"",
-                $"throw r7",
-
-                $"@receiveHandler: load r7 \"test received handler exception\"",
-                $"throw r7",
-
-                $"@burnHandler: load r7 \"test burn handler exception\"",
-                $"throw r7",
-
-                $"@OnMint: load r11 0x{addressStr}",
-                $"push r11",
-                $@"extcall ""Address()""",
-                $"pop r11",
-
-                $"load r10, {(int)EventKind.Custom}",
-                $@"load r12, ""{message}""",
-
-                $"push r12",
-                $"push r11",
-                $"push r10",
-                $@"extcall ""Runtime.Notify""",
-                "ret",
-
-                $"@return: ret",
-            };
-
-            var script = AssemblerUtils.BuildScript(scriptString, null, out var debugInfo, out var labels);
-
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, $"{symbol}Token", 1000000000, 3, flags, script, labels);
-            simulator.EndBlock();
-
-            simulator.BeginBlock();
-            var tx = simulator.MintTokens(owner, owner.Address, symbol, 1000);
-            simulator.EndBlock();
-
-            var token = simulator.Nexus.GetTokenInfo(simulator.Nexus.RootStorage, symbol);
-            var balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, owner.Address);
-            Assert.IsTrue(balance == 1000);
-
-            var events = simulator.Nexus.FindBlockByTransaction(tx).GetEventsForTransaction(tx.Hash);
-            Assert.IsTrue(events.Count(x => x.Kind == EventKind.Custom) == 1);
-
-            var eventData = events.First(x => x.Kind == EventKind.Custom).Data;
-            var eventMessage = (VMObject)Serialization.Unserialize(eventData, typeof(VMObject));
-
-            Assert.IsTrue(eventMessage.AsString() == message);
-
-            /*Assert.ThrowsException<ChainException>(() =>
-            {
-                simulator.BeginBlock();
-                simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain, symbol, 10000);
-                simulator.EndBlock();
-            });*/
-
-            balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, owner.Address);
-            Assert.IsTrue(balance == 1000);
-        }
-
-        [TestMethod]
-        public void AccountTriggersAllowance()
-        {
-            string[] scriptString;
-
-            var owner = PhantasmaKeys.Generate();
-            var target = PhantasmaKeys.Generate();
-            var other = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var symbol = "SOUL";
-            var addressStr = Base16.Encode(other.Address.ToByteArray());
-
-            scriptString = new string[]
-            {
-                $"alias r1, $temp",
-                $"alias r2, $from",
-                $"alias r3, $to",
-                $"alias r4, $symbol",
-                $"alias r5, $amount",
-
-                $"jmp @end",
-
-                $"@OnReceive: nop",
-                $"pop $from",
-                $"pop $to",
-                $"pop $symbol",
-                $"pop $amount",
-
-                $@"load r1 ""{symbol}""",
-                $@"equal r1, $symbol, $temp",
-                $"jmpnot $temp, @end",
-
-                $"load $temp 2",
-                $"div $amount $temp $temp",
-
-                $"push $temp",
-                $"push $symbol",
-                $"load r11 0x{addressStr}",
-                $"push r11",
-                $@"extcall ""Address()""",
-                $"push $to",
-
-                $"load r0 \"Runtime.TransferTokens\"",
-                $"extcall r0",
-                $"jmp @end",
-
-                $"@end: ret"
-            };
-
-
-            Dictionary<string, int> labels;
-            DebugInfo debugInfo;
-            var script = AssemblerUtils.BuildScript(scriptString, "test", out debugInfo, out labels);
-
-            var methods = TokenUtils.GetTriggersForABI(labels);
-            var abi = new ContractInterface(methods, Enumerable.Empty<ContractEvent>());
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, "KCAL", UnitConversion.GetUnitValue(DomainSettings.FuelTokenDecimals));
-            simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, "SOUL", UnitConversion.GetUnitValue(DomainSettings.StakingTokenDecimals));
-            simulator.GenerateCustomTransaction(target, ProofOfWork.None,
-                () => ScriptUtils.BeginScript()
-                        .AllowGas(target.Address, Address.Null, 1, TestGasLimit)
-                    .CallContract(NativeContractKind.Stake, nameof(StakeContract.Stake), target.Address, UnitConversion.GetUnitValue(DomainSettings.StakingTokenDecimals))
-                        .SpendGas(target.Address)
-                    .EndScript());
-            simulator.EndBlock();
-
-
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(target, ProofOfWork.None,
-                () => ScriptUtils.BeginScript()
-                        .AllowGas(target.Address, Address.Null, 1, TestGasLimit)
-                    .CallContract(NativeContractKind.Account, nameof(AccountContract.RegisterScript), target.Address, script, abi.ToByteArray())
-                        .SpendGas(target.Address)
-                    .EndScript());
-            simulator.EndBlock();
-
-            var initialBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, symbol, target.Address);
-
-            var amount = UnitConversion.ToBigInteger(5, DomainSettings.StakingTokenDecimals);
-
-            simulator.BeginBlock();
-            var tx = simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, symbol, amount * 2);
-            simulator.EndBlock();
-
-            var balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, symbol, target.Address);
-            var expectedBalance = initialBalance + amount;
-            Assert.IsTrue(balance == expectedBalance);
-
-            balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, symbol, other.Address);
-            Assert.IsTrue(balance == amount);
-        }
-
-        [TestMethod]
-        [Ignore]
-        public void AccountTriggersEventPropagation()
-        {
-            string[] scriptString;
-
-            var owner = PhantasmaKeys.Generate();
-            var target = PhantasmaKeys.Generate();
-            var symbol = "TEST";
-            var flags = TokenFlags.Transferable | TokenFlags.Finite | TokenFlags.Fungible | TokenFlags.Divisible;
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            string message = "customEvent";
-            var addressStr = Base16.Encode(target.Address.ToByteArray());
-            var isTrue = true;
-
-            scriptString = new string[]
-            {
-                $"alias r1, $triggerSend",
-                $"alias r2, $triggerReceive",
-                $"alias r3, $triggerBurn",
-                $"alias r4, $triggerMint",
-                $"alias r5, $currentTrigger",
-                $"alias r6, $comparisonResult",
-                $"alias r7, $triggerWitness",
-                $"alias r8, $currentAddress",
-                $"alias r9, $sourceAddress",
-
-                $@"load $triggerSend, ""{AccountTrigger.OnSend}""",
-                $@"load $triggerReceive, ""{AccountTrigger.OnReceive}""",
-                $@"load $triggerBurn, ""{AccountTrigger.OnBurn}""",
-                $@"load $triggerMint, ""{AccountTrigger.OnMint}""",
-                $@"load $triggerWitness, ""{AccountTrigger.OnWitness}""",
-                $"pop $currentTrigger",
-                $"pop $currentAddress",
-
-                $"equal $triggerWitness, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @witnessHandler",
-
-                $"equal $triggerSend, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @sendHandler",
-
-                $"equal $triggerReceive, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @receiveHandler",
-
-                $"equal $triggerBurn, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @burnHandler",
-
-                $"equal $triggerMint, $currentTrigger, $comparisonResult",
-                $"jmpif $comparisonResult, @OnMint",
-
-                $"jmp @end",
-
-                $"@witnessHandler: ",
-                $"load r11 0x{addressStr}",
-                $"push r11",
-                "extcall \"Address()\"",
-                $"pop $sourceAddress",
-                $"equal $sourceAddress, $currentAddress, $comparisonResult",
-                "jmpif $comparisonResult, @endWitness",
-                $"load r1 \"test witness handler xception\"",
-                $"throw r1",
-                
-                "jmp @end",
-
-                $"@sendHandler: jmp @end",
-
-                $"@receiveHandler: jmp @end",
-
-                $"@burnHandler: jmp @end",
-
-                $"@OnMint: load r11 0x{addressStr}",
-                $"push r11",
-                $@"extcall ""Address()""",
-                $"pop r11",
-
-                $"load r10, {(int)EventKind.Custom}",
-                $@"load r12, ""{message}""",
-
-                $"push r12",
-                $"push r11",
-                $"push r10",
-                $@"extcall ""Runtime.Notify""",
-
-                $"@endWitness: ret",
-                $"load r11 {isTrue}",
-                $"push r11",
-
-                $"@end: ret"
-            };
-
-            var script = AssemblerUtils.BuildScript(scriptString, null, out var something, out var labels);
-            var methods = new List<ContractMethod>();
-            methods.Add(new ContractMethod("OnMint", VMType.None, 205, new ContractParameter[0]));
-            var abi = new ContractInterface(methods, new List<ContractEvent>());
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, "KCAL", 60000000000000);
-            simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, "SOUL", UnitConversion.GetUnitValue(DomainSettings.StakingTokenDecimals)*50000);
-            simulator.GenerateCustomTransaction(target, ProofOfWork.None,
-                () => ScriptUtils.BeginScript().AllowGas(target.Address, Address.Null, 1, TestGasLimit)
-                    .CallContract("stake", "Stake", target.Address, UnitConversion.GetUnitValue(DomainSettings.StakingTokenDecimals)*50000)
-                    .CallContract("account", "RegisterScript", target.Address, script, abi.ToByteArray()).SpendGas(target.Address)
-                    .EndScript());
-            simulator.EndBlock();
-
-            simulator.BeginBlock();
-            simulator.GenerateToken(target, symbol, $"{symbol}Token", 1000000000, 3, flags);
-            var tx = simulator.MintTokens(target, target.Address, symbol, 1000);
-            simulator.EndBlock();
-
-            var accountScript = simulator.Nexus.LookUpAddressScript(simulator.Nexus.RootStorage, target.Address);
-            Assert.IsTrue(accountScript != null && accountScript.Length > 0);
-
-            var token = simulator.Nexus.GetTokenInfo(simulator.Nexus.RootStorage, symbol);
-            var balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, target.Address);
-            Assert.IsTrue(balance == 1000);
-
-            var events = simulator.Nexus.FindBlockByTransaction(tx).GetEventsForTransaction(tx.Hash);
-            Assert.IsTrue(events.Count(x => x.Kind == EventKind.Custom) == 1);
-
-            var eventData = events.First(x => x.Kind == EventKind.Custom).Data;
-            var eventMessage = (VMObject)Serialization.Unserialize(eventData, typeof(VMObject));
-
-            Assert.IsTrue(eventMessage.AsString() == message);
-            Assert.ThrowsException<ChainException>(() =>
-            {
-                simulator.BeginBlock();
-                simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain as Chain, symbol, 10);
-                simulator.EndBlock();
-            });
-
-            balance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, target.Address);
-            Assert.IsTrue(balance == 1000);
-        }
+    
 
         #region RegisterOps
