+++ Phantasma.Tests/ChainTests.cs
@@ -108,51 +108,7 @@ namespace Phantasma.LegacyTests
             Assert.IsTrue(txCount > 0);*/
         }
 
-        [TestMethod]
-        public void FuelTokenTransfer()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var accountChain = nexus.GetChainByName("account");
-            var symbol = DomainSettings.FuelTokenSymbol;
-            var token = nexus.GetTokenInfo(nexus.RootStorage, symbol);
-
-            var testUserA = PhantasmaKeys.Generate();
-            var testUserB = PhantasmaKeys.Generate();
-
-            var amount = UnitConversion.ToBigInteger(2, token.Decimals);
-
-            // Send from Genesis address to test user A
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain as Chain, symbol, amount);
-            simulator.EndBlock();
-
-            var oldBalance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUserA.Address);
-
-            Assert.IsTrue(oldBalance == amount);
-
-            // Send from test user A address to test user B
-            amount /= 2;
-            simulator.BeginBlock();
-            var tx = simulator.GenerateTransfer(testUserA, testUserB.Address, nexus.RootChain as Chain, symbol, amount);
-            simulator.EndBlock();
-
-            // verify test user balance
-            var transferBalance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUserB.Address);
-            Assert.IsTrue(transferBalance == amount);
-
-            var newBalance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUserA.Address);
-            var gasFee = nexus.RootChain.GetTransactionFee(tx);
-
-            var expectedFee = oldBalance - (newBalance + transferBalance);
-            Assert.IsTrue(expectedFee == gasFee);
-
-            var sum = transferBalance + newBalance + gasFee;
-            Assert.IsTrue(sum == oldBalance);
-        }
+        
 
         [TestMethod]
         public void CreateToken()
@@ -267,172 +223,7 @@ namespace Phantasma.LegacyTests
 
             Assert.IsTrue(transferBalance + newBalance == oldBalance);
         }
-
-        [TestMethod]
-        public void AccountRegister()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var fuelSymbol = DomainSettings.FuelTokenSymbol;
-
-            Func<PhantasmaKeys, string, bool> registerName = (keypair, name) =>
-            {
-                bool result = true;
-
-                try
-                {
-                    simulator.BeginBlock();
-                    var tx = simulator.GenerateAccountRegistration(keypair, name);
-                    var lastBlock = simulator.EndBlock().FirstOrDefault();
-
-                    if (lastBlock != null)
-                    {
-                        Assert.IsTrue(tx != null);
-
-                        var state = lastBlock.GetStateForTransaction(tx.Hash);
-                        Assert.IsTrue(state == ExecutionState.Halt);
-
-                        var evts = lastBlock.GetEventsForTransaction(tx.Hash);
-                        Assert.IsTrue(evts.Any(x => x.Kind == EventKind.AddressRegister));
-                    }
-                }
-                catch (Exception)
-                {
-                    result = false;
-                }
-
-                return result;
-            };
-
-            var testUser = PhantasmaKeys.Generate();
-
-            var token = nexus.GetTokenInfo(nexus.RootStorage, fuelSymbol);
-            var amount = UnitConversion.ToBigInteger(10, token.Decimals);
-
-            var stakeAmount = UnitConversion.ToBigInteger(3, DomainSettings.StakingTokenDecimals);
-
-            var initialBalance = nexus.RootChain.GetTokenBalance(nexus.RootStorage, DomainSettings.StakingTokenSymbol, owner.Address);
-            Assert.IsTrue(initialBalance >= stakeAmount);
-
-            // Send from Genesis address to test user
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, fuelSymbol, amount);
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, DomainSettings.StakingTokenSymbol, stakeAmount);
-            simulator.EndBlock();
-
-            // verify test user balance
-            var balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUser.Address);
-            Assert.IsTrue(balance == amount);
-
-            // make user stake enough to register a name
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(testUser, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().
-                    AllowGas(testUser.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
-                    CallContract(NativeContractKind.Stake, "Stake", testUser.Address, stakeAmount).
-                    SpendGas(testUser.Address).
-                    EndScript());
-            simulator.EndBlock();
-
-            var targetName = "hello";
-            Assert.IsTrue(targetName == targetName.ToLower());
-
-            Assert.IsFalse(registerName(testUser, targetName.Substring(3)));
-            Assert.IsFalse(registerName(testUser, targetName.ToUpper()));
-            Assert.IsFalse(registerName(testUser, targetName + "!"));
-            Assert.IsTrue(registerName(testUser, targetName));
-
-            var currentName = nexus.RootChain.GetNameFromAddress(nexus.RootStorage, testUser.Address);
-            Assert.IsTrue(currentName == targetName);
-
-            var someAddress = nexus.LookUpName(nexus.RootStorage, targetName);
-            Assert.IsTrue(someAddress == testUser.Address);
-
-            Assert.IsFalse(registerName(testUser, "other"));
-        }
-
-        [TestMethod]
-        public void AccountMigrate()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var symbol = DomainSettings.FuelTokenSymbol;
-
-            var testUser = PhantasmaKeys.Generate();
-
-            var token = nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            var amount = UnitConversion.ToBigInteger(10, token.Decimals);
-
-            var stakeAmount = UnitConversion.ToBigInteger(3, DomainSettings.StakingTokenDecimals);
-
-            // Send from Genesis address to test user
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, symbol, amount);
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, DomainSettings.StakingTokenSymbol, stakeAmount);
-            simulator.EndBlock();
-
-            // verify test user balance
-            var balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUser.Address);
-            Assert.IsTrue(balance == amount);
-
-            // make user stake enough to register a name
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(testUser, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().
-                    AllowGas(testUser.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
-                    CallContract(NativeContractKind.Stake, "Stake", testUser.Address, stakeAmount).
-                    SpendGas(testUser.Address).
-                    EndScript());
-            simulator.EndBlock();
-
-            var targetName = "hello";
-            Assert.IsTrue(targetName == targetName.ToLower());
-
-            simulator.BeginBlock();
-            var tx = simulator.GenerateAccountRegistration(testUser, targetName);
-            var lastBlock = simulator.EndBlock().FirstOrDefault();
-
-            if (lastBlock != null)
-            {
-                Assert.IsTrue(tx != null);
-
-                var evts = lastBlock.GetEventsForTransaction(tx.Hash);
-                Assert.IsTrue(evts.Any(x => x.Kind == EventKind.AddressRegister));
-            }
-
-
-            var currentName = nexus.RootChain.GetNameFromAddress(nexus.RootStorage, testUser.Address);
-            Assert.IsTrue(currentName == targetName);
-
-            var someAddress = nexus.LookUpName(nexus.RootStorage, targetName);
-            Assert.IsTrue(someAddress == testUser.Address);
-
-            var migratedUser = PhantasmaKeys.Generate();
-
-            simulator.BeginBlock();
-            tx = simulator.GenerateCustomTransaction(testUser, ProofOfWork.None, () =>
-            {
-                return ScriptUtils.BeginScript().
-                     AllowGas(testUser.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
-                     CallContract(NativeContractKind.Account, nameof(AccountContract.Migrate), testUser.Address, migratedUser.Address).
-                     SpendGas(testUser.Address).
-                     EndScript();
-            });
-            simulator.EndBlock().FirstOrDefault();
-
-            currentName = nexus.RootChain.GetNameFromAddress(nexus.RootStorage, testUser.Address);
-            Assert.IsFalse(currentName == targetName);
-
-            var newName = nexus.RootChain.GetNameFromAddress(nexus.RootStorage, migratedUser.Address);
-            Assert.IsTrue(newName == targetName);
-        }
-
+        
         [TestMethod]
         public void SimpleTransfer()
         {
@@ -766,104 +557,6 @@ namespace Phantasma.LegacyTests
             Assert.IsTrue(price == 5);
         }
 
-        /*
-        [TestMethod]
-        public void GetRatesForSwap()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var script = new ScriptBuilder().CallContract("swap", "GetRates", "SOUL", UnitConversion.GetUnitValue(DomainSettings.StakingTokenDecimals)).EndScript();
-
-            var result = nexus.RootChain.InvokeScript(nexus.RootStorage, script);
-
-            var temp = result.ToObject();
-            var rates = (SwapPair[])temp;
-
-            decimal targetRate = 0;
-
-            foreach (var entry in rates)
-            {
-                if (entry.Symbol == DomainSettings.FuelTokenSymbol)
-                {
-                    targetRate = UnitConversion.ToDecimal(entry.Value, DomainSettings.FuelTokenDecimals);
-                    break;
-                }
-            }
-
-            Assert.IsTrue(targetRate == 5m);
-        }*/
-
-        [TestMethod]
-        public void TransferToAccountName()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var symbol = DomainSettings.FuelTokenSymbol;
-
-            Func<PhantasmaKeys, string, bool> registerName = (keypair, name) =>
-            {
-                bool result = true;
-
-                try
-                {
-                    simulator.BeginBlock();
-                    var tx = simulator.GenerateAccountRegistration(keypair, name);
-                    var lastBlock = simulator.EndBlock().FirstOrDefault();
-
-                    if (lastBlock != null)
-                    {
-                        Assert.IsTrue(tx != null);
-
-                        var evts = lastBlock.GetEventsForTransaction(tx.Hash);
-                        Assert.IsTrue(evts.Any(x => x.Kind == EventKind.AddressRegister));
-                    }
-                }
-                catch (Exception)
-                {
-                    result = false;
-                }
-
-                return result;
-            };
-
-            var targetName = "hello";
-            var testUser = PhantasmaKeys.Generate();
-            var token = nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            var amount = UnitConversion.ToBigInteger(10, token.Decimals);
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, symbol, amount);
-            simulator.GenerateTransfer(owner, testUser.Address, nexus.RootChain, DomainSettings.StakingTokenSymbol, amount);
-            simulator.GenerateCustomTransaction(testUser, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().AllowGas(testUser.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit)
-                    .CallContract(NativeContractKind.Stake, "Stake", testUser.Address, amount).
-                    SpendGas(testUser.Address).EndScript());
-            simulator.EndBlock();
-
-            Assert.IsTrue(registerName(testUser, targetName));
-
-            // Send from Genesis address to test user
-            var transferAmount = 1;
-
-            var initialFuelBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUser.Address);
-
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(owner, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().AllowGas(owner.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit)
-                    .TransferTokens(token.Symbol, owner.Address, targetName, transferAmount)
-                    .SpendGas(owner.Address).EndScript());
-            simulator.EndBlock();
-
-            var finalFuelBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, testUser.Address);
-
-            Assert.IsTrue(finalFuelBalance - initialFuelBalance == transferAmount);
-        }
 
         [TestMethod]
         public void SideChainTransferDifferentAccounts()
@@ -1081,428 +774,7 @@ namespace Phantasma.LegacyTests
             // TODO  verify balances
         }
 
-        [TestMethod]
-        public void NftMint()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var chain = nexus.RootChain;
-
-            var symbol = "COOL";
-
-            var testUser = PhantasmaKeys.Generate();
-
-            // Create the token CoolToken as an NFT
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, "CoolToken", 0, 0, TokenFlags.Transferable);
-            simulator.EndBlock();
-
-            var token = simulator.Nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            Assert.IsTrue(nexus.TokenExists(nexus.RootStorage, symbol), "Can't find the token symbol");
-
-            // verify nft presence on the user pre-mint
-            var ownerships = new OwnershipSheet(symbol);
-            var ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the sender already have a CoolToken?");
-
-            var tokenROM = new byte[] { 0x1, 0x3, 0x3, 0x7 };
-            var tokenRAM = new byte[] { 0x1, 0x4, 0x4, 0x6 };
-
-            // Mint a new CoolToken to test address
-            simulator.BeginBlock();
-            simulator.MintNonFungibleToken(owner, testUser.Address, symbol, tokenROM, tokenRAM, 0);
-            simulator.EndBlock();
-
-            // obtain tokenID
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-
-            // verify nft presence on the user post-mint
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-
-            // check used storage
-            var tokenAddress = TokenUtils.GetContractAddress(symbol);
-            var usedStorage = (int)nexus.RootChain.InvokeContract(nexus.RootChain.Storage, "storage", nameof(StorageContract.GetUsedSpace), tokenAddress).AsNumber();
-            var minExpectedSize = tokenROM.Length + tokenRAM.Length;
-            Assert.IsTrue(usedStorage >= minExpectedSize);
-
-            //verify that the present nft is the same we actually tried to create
-            var tokenID = ownedTokenList.First();
-            var nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenID);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) && nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-
-            var currentSupply = chain.GetTokenSupply(chain.Storage, symbol);
-            Assert.IsTrue(currentSupply == 1, "why supply did not increase?");
-
-            var testScript = new ScriptBuilder().CallNFT(symbol, 0, "getName", tokenID).EndScript();
-            var temp  = simulator.Nexus.RootChain.InvokeScript(simulator.Nexus.RootStorage, testScript);
-            var testResult = temp.AsString();
-            Assert.IsTrue(testResult == "CoolToken");
-        }
-
-
-        [TestMethod]
-        public void NftBurn()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var chain = nexus.RootChain;
-
-            var symbol = "COOL";
-
-            var testUser = PhantasmaKeys.Generate();
-
-            BigInteger seriesID = 123;
-
-            // Create the token CoolToken as an NFT
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, "CoolToken", 0, 0, TokenFlags.Burnable, null, null, null, (uint)seriesID);
-            simulator.EndBlock();
-
-            var series = nexus.GetTokenSeries(nexus.RootStorage, symbol, seriesID);
-
-            Assert.IsTrue(series.MintCount == 0, "nothing should be minted yet");
-
-            // Send some KCAL and SOUL to the test user (required for gas used in "burn" transaction)
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUser.Address, chain, DomainSettings.FuelTokenSymbol, UnitConversion.ToBigInteger(1, DomainSettings.FuelTokenDecimals));
-            simulator.GenerateTransfer(owner, testUser.Address, chain, DomainSettings.StakingTokenSymbol, UnitConversion.ToBigInteger(1, DomainSettings.StakingTokenDecimals));
-            simulator.EndBlock();
-
-            var token = simulator.Nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            Assert.IsTrue(nexus.TokenExists(nexus.RootStorage, symbol), "Can't find the token symbol");
-
-            // verify nft presence on the user pre-mint
-            var ownerships = new OwnershipSheet(symbol);
-            var ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the user already have a CoolToken?");
-
-            var tokenROM = new byte[] { 0x1, 0x3, 0x3, 0x7 };
-            var tokenRAM = new byte[] { 0x1, 0x4, 0x4, 0x6 };
-
-            // Mint a new CoolToken to test address
-            simulator.BeginBlock();
-            simulator.MintNonFungibleToken(owner, testUser.Address, symbol, tokenROM, tokenRAM, seriesID);
-            simulator.EndBlock();
-
-            // obtain tokenID
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-            var tokenID = ownedTokenList.First();
-
-            // verify nft presence on the user post-mint
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the user not have one now?");
-
-            var ownerAddress = ownerships.GetOwner(chain.Storage, tokenID);
-            Assert.IsTrue(ownerAddress == testUser.Address);
-
-            //verify that the present nft is the same we actually tried to create
-            var tokenId = ownedTokenList.ElementAt(0);
-            var nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) || nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-
-            Assert.IsTrue(nft.Infusion.Length == 0); // nothing should be infused yet
-
-            var infuseSymbol = DomainSettings.StakingTokenSymbol;
-            var infuseAmount = UnitConversion.ToBigInteger(1, DomainSettings.StakingTokenDecimals);
-
-            var prevBalance = nexus.RootChain.GetTokenBalance(nexus.RootChain.Storage, infuseSymbol, testUser.Address);
-
-            // Infuse some KCAL to the CoolToken
-            simulator.BeginBlock();
-            simulator.InfuseNonFungibleToken(testUser, symbol, tokenId, infuseSymbol, infuseAmount);
-            simulator.EndBlock();
-
-            nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.Infusion.Length == 1); // should have something infused now
-
-            var infusedBalance = nexus.RootChain.GetTokenBalance(nexus.RootChain.Storage, infuseSymbol, DomainSettings.InfusionAddress);
-            Assert.IsTrue(infusedBalance == infuseAmount); // should match
-
-            var curBalance = nexus.RootChain.GetTokenBalance(nexus.RootChain.Storage, infuseSymbol, testUser.Address);
-            Assert.IsTrue(curBalance + infusedBalance == prevBalance); // should match
-
-            prevBalance = curBalance;
-
-            // burn the token
-            simulator.BeginBlock();
-            simulator.GenerateNftBurn(testUser, chain, symbol, tokenId);
-            simulator.EndBlock();
-
-            //verify the user no longer has the token
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the user still have it post-burn?");
-
-            // verify that the user received the infused assets
-            curBalance = nexus.RootChain.GetTokenBalance(nexus.RootChain.Storage, infuseSymbol, testUser.Address);
-            Assert.IsTrue(curBalance == prevBalance + infusedBalance); // should match
-
-            var burnedSupply = nexus.GetBurnedTokenSupply(nexus.RootStorage, symbol);
-            Assert.IsTrue(burnedSupply == 1);
-
-            var burnedSeriesSupply = nexus.GetBurnedTokenSupplyForSeries(nexus.RootStorage, symbol, seriesID);
-            Assert.IsTrue(burnedSeriesSupply == 1);
-        }
-
-        [TestMethod]
-        public void NftTransfer()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var chain = nexus.RootChain;
-
-            var nftKey = PhantasmaKeys.Generate();
-            var symbol = "COOL";
-            var nftName = "CoolToken";
-
-            var sender = PhantasmaKeys.Generate();
-            var receiver = PhantasmaKeys.Generate();
-
-            // Send some SOUL to the test user (required for gas used in "transfer" transaction)
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, sender.Address, chain, DomainSettings.FuelTokenSymbol, UnitConversion.ToBigInteger(1, DomainSettings.FuelTokenDecimals));
-            simulator.EndBlock();
-
-            // Create the token CoolToken as an NFT
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, nftName, 0, 0, TokenFlags.Transferable);
-            simulator.EndBlock();
-
-            var token = simulator.Nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            Assert.IsTrue(nexus.TokenExists(nexus.RootStorage, symbol), "Can't find the token symbol");
-
-            // verify nft presence on the sender pre-mint
-            var ownerships = new OwnershipSheet(symbol);
-            var ownedTokenList = ownerships.Get(chain.Storage, sender.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the sender already have a CoolToken?");
-
-            var tokenROM = new byte[] { 0x1, 0x3, 0x3, 0x7 };
-            var tokenRAM = new byte[] { 0x1, 0x4, 0x4, 0x6 };
-
-            // Mint a new CoolToken 
-            simulator.BeginBlock();
-            simulator.MintNonFungibleToken(owner, sender.Address, symbol, tokenROM, tokenRAM, 0);
-            simulator.EndBlock();
-
-            // obtain tokenID
-            ownedTokenList = ownerships.Get(chain.Storage, sender.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-            var tokenID = ownedTokenList.First();
-
-            // verify nft presence on the sender post-mint
-            ownedTokenList = ownerships.Get(chain.Storage, sender.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-
-            //verify that the present nft is the same we actually tried to create
-            var tokenId = ownedTokenList.ElementAt(0);
-            var nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) || nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-
-            // verify nft presence on the receiver pre-transfer
-            ownedTokenList = ownerships.Get(chain.Storage, receiver.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the receiver already have a CoolToken?");
-
-            // transfer that nft from sender to receiver
-            simulator.BeginBlock();
-            var txA = simulator.GenerateNftTransfer(sender, receiver.Address, chain, symbol, tokenId);
-            simulator.EndBlock();
-
-            // verify nft presence on the receiver post-transfer
-            ownedTokenList = ownerships.Get(chain.Storage, receiver.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the receiver not have one now?");
-
-            //verify that the transfered nft is the same we actually tried to create
-            tokenId = ownedTokenList.ElementAt(0);
-            nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) || nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-        }
-
-        [TestMethod]
-        public void NftMassMint()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var chain = nexus.RootChain;
-
-            var symbol = "COOL";
-            
-            var testUser = PhantasmaKeys.Generate();
-
-            // Create the token CoolToken as an NFT
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, "CoolToken", 0, 0, TokenFlags.Transferable);
-            simulator.EndBlock();
-
-            var tokenAddress = TokenUtils.GetContractAddress(symbol);
-            var storageStakeAmount = UnitConversion.ToBigInteger(100000, DomainSettings.StakingTokenDecimals);
-
-            // Add some storage to the NFT contract
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, tokenAddress, chain, DomainSettings.StakingTokenSymbol, storageStakeAmount);
-
-            simulator.GenerateCustomTransaction(owner, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().
-                    AllowGas(owner.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
-                    CallContract(NativeContractKind.Stake, nameof(StakeContract.Stake), tokenAddress, storageStakeAmount).
-                    SpendGas(owner.Address).
-                    EndScript());
-            simulator.EndBlock();
-
-
-            Assert.IsTrue(nexus.TokenExists(nexus.RootStorage, symbol), "Can't find the token symbol");
-
-            // verify nft presence on the user pre-mint
-            var ownerships = new OwnershipSheet(symbol);
-            var ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the sender already have nfts?");
-
-            var tokenRAM = new byte[] { 0x1, 0x4, 0x4, 0x6 };
-
-            var nftCount = 1000;
-
-            var initialKCAL = nexus.RootChain.GetTokenBalance(nexus.RootStorage, DomainSettings.FuelTokenSymbol, owner.Address);
-
-            // Mint several nfts to test limit per tx
-            simulator.BeginBlock();
-            for (int i=1; i<=nftCount; i++)
-            {
-                var tokenROM = BitConverter.GetBytes(i);
-                simulator.MintNonFungibleToken(owner, testUser.Address, symbol, tokenROM, tokenRAM, 0);
-            }
-            var block = simulator.EndBlock().First();
-
-            Assert.IsTrue(block.TransactionCount == nftCount);
-
-            // obtain tokenID
-            ownedTokenList = ownerships.Get(chain.Storage, testUser.Address);
-            var ownedTotal = ownedTokenList.Count();
-            Assert.IsTrue(ownedTotal == nftCount);
-
-            var currentKCAL = nexus.RootChain.GetTokenBalance(nexus.RootStorage, DomainSettings.FuelTokenSymbol, owner.Address);
-
-            var fee = initialKCAL - currentKCAL;
-
-            var convertedFee = UnitConversion.ToDecimal(fee, DomainSettings.FuelTokenDecimals);
-
-            Assert.IsTrue(fee > 0);
-        }
-
-        [TestMethod]
-        [Ignore] //TODO side chain transfers of NFTs do currently not work, because Storage contract is not deployed on the side chain.
-        public void SidechainNftTransfer()
-        {
-            var owner = PhantasmaKeys.Generate();
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var sourceChain = nexus.RootChain;
-
-            var symbol = "COOL";
-
-            var sender = PhantasmaKeys.Generate();
-            var receiver = PhantasmaKeys.Generate();
-
-            var fullAmount = UnitConversion.ToBigInteger(10, DomainSettings.FuelTokenDecimals);
-            var smallAmount = fullAmount / 2;
-            Assert.IsTrue(smallAmount > 0);
-
-            // Send some SOUL to the test user (required for gas used in "transfer" transaction)
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, sender.Address, sourceChain, DomainSettings.FuelTokenSymbol, fullAmount);
-            simulator.GenerateChain(owner, DomainSettings.ValidatorsOrganizationName, "main", "test");
-            simulator.EndBlock();
-
-            var targetChain = nexus.GetChainByName("test");
-
-            // Create the token CoolToken as an NFT
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, symbol, "CoolToken", 0, 0, TokenFlags.Transferable);
-            simulator.EndBlock();
-
-            var token = simulator.Nexus.GetTokenInfo(nexus.RootStorage, symbol);
-            Assert.IsTrue(nexus.TokenExists(nexus.RootStorage, symbol), "Can't find the token symbol");
-
-            // verify nft presence on the sender pre-mint
-            var ownerships = new OwnershipSheet(symbol);
-            var ownedTokenList = ownerships.Get(sourceChain.Storage, sender.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the sender already have a CoolToken?");
-
-            var tokenROM = new byte[] { 0x1, 0x3, 0x3, 0x7 };
-            var tokenRAM = new byte[] { 0x1, 0x4, 0x4, 0x6 };
-
-            // Mint a new CoolToken 
-            simulator.BeginBlock();
-            simulator.MintNonFungibleToken(owner, sender.Address, symbol, tokenROM, tokenRAM, 0);
-            simulator.EndBlock();
-
-            // obtain tokenID
-            ownedTokenList = ownerships.Get(sourceChain.Storage, sender.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-            var tokenID = ownedTokenList.First();
-
-            // verify nft presence on the sender post-mint
-            ownedTokenList = ownerships.Get(sourceChain.Storage, sender.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the sender not have one now?");
-
-            //verify that the present nft is the same we actually tried to create
-            var tokenId = ownedTokenList.ElementAt(0);
-            var nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) || nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-
-            // verify nft presence on the receiver pre-transfer
-            ownedTokenList = ownerships.Get(targetChain.Storage, receiver.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the receiver already have a CoolToken?");
-
-            var extraFee = UnitConversion.ToBigInteger(0.001m, DomainSettings.FuelTokenDecimals);
-
-            // transfer that nft from sender to receiver
-            simulator.BeginBlock();
-            var txA = simulator.GenerateSideChainSend(sender, symbol, sourceChain, receiver.Address, targetChain, tokenId, extraFee);
-            simulator.EndBlock();
-
-            var blockAHash = nexus.RootChain.GetLastBlockHash();
-            var blockA = nexus.RootChain.GetBlockByHash(blockAHash);
-
-            Console.WriteLine("step 1");
-            // finish the chain transfer
-            simulator.BeginBlock();
-            simulator.GenerateSideChainSettlement(receiver, nexus.RootChain, targetChain, txA);
-            Assert.IsTrue(simulator.EndBlock().Any());
-
-            // verify the sender no longer has it
-            ownedTokenList = ownerships.Get(sourceChain.Storage, sender.Address);
-            Assert.IsTrue(!ownedTokenList.Any(), "How does the sender still have one?");
-
-            // verify nft presence on the receiver post-transfer
-            ownedTokenList = ownerships.Get(targetChain.Storage, receiver.Address);
-            Assert.IsTrue(ownedTokenList.Count() == 1, "How does the receiver not have one now?");
-
-            //verify that the transfered nft is the same we actually tried to create
-            tokenId = ownedTokenList.ElementAt(0);
-            nft = nexus.ReadNFT(nexus.RootStorage, symbol, tokenId);
-            Assert.IsTrue(nft.ROM.SequenceEqual(tokenROM) || nft.RAM.SequenceEqual(tokenRAM),
-                "And why is this NFT different than expected? Not the same data");
-        }
+ 
 
         [TestMethod]
         public void NoGasSameChainTransfer()
