+++ Phantasma.Tests/Phantasma.Tests.csproj
@@ -14,6 +14,7 @@
     <PackageReference Include="MSTest.TestAdapter" Version="2.2.3" />
     <PackageReference Include="MSTest.TestFramework" Version="2.2.3" />
     <PackageReference Include="Neo" Version="2.12.2" />
+    <PackageReference Include="xunit" Version="2.4.2" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Phantasma.Business\src\Phantasma.Business.csproj" />
